tests: remove use of transmute
This is obviously UB, as I was cast a u32 surrogate to a char.

I'm not sure why I wrote this code. I did write it before
char::from_u32_unchecked was stabilized, and maybe the story around UB
and type validity invariants wasn't as clear to me back then. Either
way, we work around it by just hand-rolling UTF-8 encoding.

This fix is ported from a similar fix I made to the tests when I copied
the code into the regex-syntax crate.

Fixes #13